backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
python-dotenv
google-generativeai
